# tests/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateEvaluationRequest": {
        "additionalProperties": false,
        "description": "\"Save draft\" for a new evaluation. Everything else takes the design's\ndefaults: the active template, temperature 0.0, seed 42, size `full`.",
        "properties": {
          "corpus_id": {
            "title": "Corpus Id",
            "type": "string"
          },
          "feature_config_id": {
            "title": "Feature Config Id",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "corpus_id",
          "feature_config_id"
        ],
        "title": "CreateEvaluationRequest",
        "type": "object"
      },
...
      "DiscardResponse": {
        "additionalProperties": false,
        "description": "What a discard actually removed.\n\nA body rather than a bare `204`, because the counts are the only record\nthat survives the call: nothing is written anywhere saying this happened\n(\u00a717.3), so the response *is* the receipt.",
        "properties": {
          "extractions": {
            "title": "Extractions",
            "type": "integer"
          },
          "files": {
            "title": "Files",
            "type": "integer"
          },
          "forced": {
            "default": false,
            "title": "Forced",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "mismatches": {
            "title": "Mismatches",
            "type": "integer"
          },
          "runs": {
            "title": "Runs",
            "type": "integer"
          },
          "scores": {
            "title": "Scores",
            "type": "integer"
          },
          "tagged_mismatches": {
            "title": "Tagged Mismatches",
            "type": "integer"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id",
          "runs",
          "extractions",
          "scores",
          "mismatches",
          "tagged_mismatches",
          "files"
        ],
        "title": "DiscardResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The body of every 4xx that is not a validation error.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvaluationDraftResponse": {
        "additionalProperties": false,
        "description": "An evaluation's setup \u2014 the six numbered steps, as data.",
        "properties": {
          "corpus_id": {
            "title": "Corpus Id",
            "type": "string"
          },
          "evaluation_id": {
            "title": "Evaluation Id",
            "type": "string"
          },
          "feature_config_id": {
            "title": "Feature Config Id",
            "type": "string"
          },
          "launched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt_language": {
            "default": "de",
            "title": "Prompt Language",
            "type": "string"
          },
          "prompt_template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Template Id"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          },
          "selected_models": {
            "items": {
              "type": "string"
            },
            "title": "Selected Models",
            "type": "array"
          },
          "size": {
            "$ref": "#/components/schemas/EvaluationSize",
            "default": "full"
          },
          "temperature": {
            "default": 0.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "evaluation_id",
          "name",
          "corpus_id",
          "feature_config_id"
        ],
        "title": "EvaluationDraftResponse",
        "type": "object"
      },
      "EvaluationLaunchResponse": {
        "additionalProperties": false,
        "description": "The launch commit, plus the task the worker runs under.\n\nTwo things in one response because they happen together and the view\nneeds both: the pinned evaluation to render, and the task id to poll.",
        "properties": {
          "evaluation": {
            "$ref": "#/components/schemas/EvaluationResponse"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "evaluation",
          "task_id"
        ],
        "title": "EvaluationLaunchResponse",
        "type": "object"
      },
      "EvaluationResponse": {
        "additionalProperties": false,
        "description": "One whole Evaluation screen.",
        "properties": {
          "can_launch": {
            "default": false,
            "title": "Can Launch",
            "type": "boolean"
          },
          "connection": {
            "$ref": "#/components/schemas/ConnectionResponse"
          },
          "corpus_label": {
            "default": "",
            "title": "Corpus Label",
            "type": "string"
          },
          "corpus_record_count": {
            "default": 0,
            "title": "Corpus Record Count",
            "type": "integer"
          },
          "dev_record_max": {
            "default": 0,
            "title": "Dev Record Max",
            "type": "integer"
          },
          "draft": {
            "$ref": "#/components/schemas/EvaluationDraftResponse"
          },
          "feature_config_label": {
            "default": "",
            "title": "Feature Config Label",
            "type": "string"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelChoiceResponse"
            },
            "title": "Models",
            "type": "array"
          },
          "progress": {
            "items": {
              "$ref": "#/components/schemas/RunProgressResponse"
            },
            "title": "Progress",
            "type": "array"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "runs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunPage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "draft",
          "connection"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
      "EvaluationSize": {
        "description": "The design's step 6: \"Evaluation \u00b7 all N\" or \"Dev \u00b7 N records\".\n\nA `DEV` selection takes the **first `RA2_DEV_RECORD_MAX` records by id** \u2014\ndeterministic, because \"a re-run is a check, not a new sample\" is false the\nmoment the record selection is random (\u00a715 F9). The seed fixes what the\nmodel does with what it sees; determinism has to cover what it is *shown*\ntoo.",
        "enum": [
          "full",
          "dev"
        ],
        "title": "EvaluationSize",
        "type": "string"
      },
...
      "FeatureValidationErrorResponse": {
        "additionalProperties": false,
        "description": "422 from a blocked freeze. **Nothing was frozen** (mvp-spec.md \u00a77/\u00a78.2).",
        "properties": {
          "detail": {
            "default": "blocking feature validation errors; the set was not frozen",
            "title": "Detail",
            "type": "string"
          },
          "validation_errors": {
            "items": {
              "type": "string"
            },
            "title": "Validation Errors",
            "type": "array"
          }
        },
        "required": [
          "validation_errors"
        ],
        "title": "FeatureValidationErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "RunPage",
        "type": "object"
      },
...
      "UpdateEvaluationRequest": {
        "additionalProperties": false,
        "description": "Edit any of the six setup steps. **Draft only** \u2014 editing a launched\nevaluation is 409.",
        "properties": {
          "corpus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Id"
          },
          "feature_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Config Id"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "prompt_language": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Language"
          },
          "prompt_template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Template Id"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "selected_models": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Models"
          },
          "size": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationSize"
              },
              {
                "type": "null"
              }
            ]
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          }
        },
        "title": "UpdateEvaluationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Road accident report analysis \u2014 import, census and evaluation.",
    "title": "RA2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluations": {
      "get": {
        "operationId": "list_evaluations_api_v1_evaluations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationDraftResponse"
                  },
                  "title": "Response List Evaluations Api V1 Evaluations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Evaluations",
        "tags": [
          "evaluations"
        ]
      },
      "post": {
        "description": "\"Save draft\" \u2014 the row exists before the inputs are final.",
        "operationId": "save_draft_api_v1_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Draft",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "delete": {
        "description": "Discard an evaluation and its runs (\u00a717.1).\n\n**The corpus, the feature config and the prompt versions it cites are\nuntouched** \u2014 they are `RESTRICT`, and discarding an evaluation is not a\nway to delete a corpus.\n\n409 when any run is `queued` or `running` (G1), and when tagged mismatches\nwould be destroyed without `force` (G2).",
        "operationId": "discard_evaluation_api_v1_evaluations__evaluation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Evaluation",
        "tags": [
          "evaluations"
        ]
      },
      "get": {
        "description": "One whole Evaluation screen: setup, models, connection, progress, runs\nand provenance.",
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "evaluations"
        ]
      },
      "put": {
        "description": "409 once `launched_at` is set \u2014 an evaluation is immutable after. 422\nwhen the edit selects a model the host is known not to have the VRAM\nfor.",
        "operationId": "update_draft_api_v1_evaluations__evaluation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Update Draft",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/evaluations/{evaluation_id}/discard-preview": {
      "get": {
        "description": "The counts summed over this evaluation's runs. Writes nothing.",
        "operationId": "discard_preview_api_v1_evaluations__evaluation_id__discard_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscardPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Preview",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/evaluations/{evaluation_id}/launch": {
      "post": {
        "description": "Pin the inputs, snapshot the codelists and fingerprints, create one\n`queued` run per selected model \u2014 then submit the worker.\n\nThe snapshot and the runs land in **one transaction**\n(`EvaluationService.launch`); submitting the work is a second step, so a\nfailed submit cannot leave a half-pinned evaluation. The returned\n`evaluation` reflects the state right after that transaction commits \u2014\nnot after the (possibly already-finished, on a synchronous task runner)\nwork \u2014 the UI polls `task_id` for progress.",
        "operationId": "launch_api_v1_evaluations__evaluation_id__launch_post",
        "parameters": [
          {
            "in": "path",
            "name": "evaluation_id",
            "required": true,
            "schema": {
              "title": "Evaluation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationLaunchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Launch",
        "tags": [
          "evaluations"
        ]
      }
    },
...
  }
}